+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [next]
 
+- feat(): Add has method to classRegistry to allow to check if a class exists. (fixes #10001)
+
 ## [6.1.0]
 
 - fix(): Avoid errors on restoring custom properties that pass the lazy detection of shadow,gradient,pattern and clipPath. [#10001](https://github.com/fabricjs/fabric.js/issues/10001)